+++ src/app/(board)/boards/page.tsx
@@ -1,32 +1,13 @@
-"use client";
-
-import { useSearchParams } from "next/navigation";
-import { Category } from "@/types/article";
-import BoardMain from "@/app/(board)/boards/components/BoardMain";
-import PaginationBar from "@/app/(board)/boards/components/PaginationBar";
-import useInitializeArticles from "@/hook/useInitializeArticles";
+import { Suspense } from "react";
 import LoadingSpinner from "@/components/LoadingSpinner";
+import BoardsContent from "@/app/(board)/boards/BoardsContent";
 
-const BoardsContent = () => {
-  if (typeof window === "undefined") return;
-
-  const searchParams = useSearchParams();
-  const page = Number(searchParams.get("page") ?? "1");
-  const category = searchParams.get("category") ?? "ALL";
-
-  // 최초 마운트 시, 전역 store에 allArticles, ArticlesByCategory를 set해줌
-  const isInitialized = useInitializeArticles();
-
-  if (!isInitialized) {
-    return <LoadingSpinner />;
-  }
-
+const Page = () => {
   return (
-    <div className="relative h-full w-full">
-      <BoardMain currentPage={page} category={category as Category} />
-      <PaginationBar currentPage={page} />
-    </div>
+    <Suspense fallback={<LoadingSpinner />}>
+      <BoardsContent />
+    </Suspense>
   );
 };
 
-export default BoardsContent;
+export default Page;
